requirements.txt
edge-tts
moviepy
Pillow
numpy
yt-dlp
